Gh0st1nSh3ll: - Разнес все как необходимо хранить
