+++ src/main/java/Elements/ButtonElement/ButtonElement.java
@@ -1,0 +1,7 @@
+package Elements.ButtonElement;
+
+public class ButtonElement {
+
+    // тут наследуемся от бейсЭлемента и просто добавляем методы для кнопки типичные
+    // С другими элементами аналогично: наследуемся от бейсЭлемента и дописываем реализацаю
+}
