+++ src/main/java/Helpers/ScreenshotUtil.java
@@ -1,0 +1,21 @@
+package Helpers;
+
+import io.qameta.allure.Allure;
+import io.qameta.allure.Attachment;
+import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
+import org.openqa.selenium.WebDriver;
+
+import java.io.ByteArrayInputStream;
+import java.io.File;
+
+public class ScreenshotUtil {
+
+    @Attachment(value = "Screenshot", type = "image/png")
+    public static byte[] takeScreenshot(WebDriver driver) {
+        byte[] screenshot = ((TakesScreenshot) driver).getScreenshotAs(OutputType.BYTES);
+        Allure.addAttachment("screenshot", new ByteArrayInputStream(screenshot));
+        return screenshot;
+
+    }
+}
