+++ src/test/java/demoqa/UiTests/NeNashelKudaZakinut/DynamicTests.java
@@ -1,0 +1,31 @@
+package demoqa.UiTests.NeNashelKudaZakinut;
+
+import BaseTest;
+import Pages.DemoqaPages.NeNashelKudaZakinut.DynamicPage;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+
+public class DynamicTests extends BaseTest {
+    DynamicPage dynamicPage = new DynamicPage();
+
+    @Test
+    public void IsColorChangedTest() throws InterruptedException {
+        dynamicPage.navigateToBasePage();
+        dynamicPage.clickDynamicPage();
+
+        dynamicPage.storeInitialColor();
+        Thread.sleep(5000);
+        Assertions.assertTrue(dynamicPage.isColorChanged());
+    }
+
+    @Test
+    public void IsButtonPresentedTest() throws InterruptedException {
+        dynamicPage.navigateToBasePage();
+        dynamicPage.clickDynamicPage();
+
+        Assertions.assertFalse(dynamicPage.isVisibleAfterButtonPresent());
+        Thread.sleep(5000);
+        Assertions.assertTrue(dynamicPage.isVisibleAfterButtonPresent());
+    }
+}
+
