--- src/test/java/demoqa/tests/FilesTests.java
@@ -1,31 +1,0 @@
-package demoqa.tests;
-
-import demoqa.base.BaseTest;
-import demoqa.pages.FilesPage;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class FilesTests extends BaseTest {
-    FilesPage filesPage = new FilesPage();
-
-    @Test
-    public void IsFileDownloadedTest(){
-        filesPage.navigateToBasePage();
-        filesPage.clickFilesPage();
-
-        filesPage.clickDownloadButton();
-        Assertions.assertTrue(filesPage.isFileDownloaded());
-
-        filesPage.uploadDownloadedFile();
-        Assertions.assertTrue(filesPage.isUploadedFilePathCorrect());
-    }
-
-//    @Test
-//    public void IsFileUploadedTest(){
-//        filesPage.navigateToBasePage();
-//        filesPage.clickFilesPage();
-//
-//        filesPage.uploadDownloadedFile();
-//        Assertions.assertTrue(filesPage.isUploadedFilePathCorrect());
-//    }
-}
